validate requst params

+++ handler/events.go
@@ -17,8 +17,17 @@ func (hdlr *handler) CreateEvent(wtr http.ResponseWriter, req *http.Request) {
 
 	var eventData dto.Event
 	json.NewDecoder(req.Body).Decode(&eventData)
-	eventData.UserId = userId
 
+	if !isDateValid(eventData.Date) ||
+		!isTimeValid(eventData.StartTime) ||
+		!isTimeValid(eventData.EndTime) ||
+		eventData.StartTime > eventData.EndTime ||
+		eventData.Content == "" {
+		throwBadReqErr(wtr, "Incorrect data")
+		return
+	}
+
+	eventData.UserId = userId
 	eventId, err := hdlr.service.CreateEvent(eventData)
 	if err != nil {
 		throwIntServerErr(wtr)
@@ -39,10 +48,19 @@ func (hdlr *handler) GetEvents(wtr http.ResponseWriter, req *http.Request) {
 
 	req.ParseForm()
 	periodParams := dto.PeriodParams{
-		PeriodStart: req.Form["periodStart"][0],
-		PeriodEnd:   req.Form["periodEnd"][0],
+		PeriodStart: req.Form.Get("periodStart"),
+		PeriodEnd:   req.Form.Get("periodEnd"),
 		UserId:      userId,
 	}
+
+	if !isDateValid(periodParams.PeriodStart) ||
+		!isDateValid(periodParams.PeriodEnd) ||
+		periodParams.PeriodStart > periodParams.PeriodEnd ||
+		periodParams.UserId == "" {
+		throwBadReqErr(wtr, "Incorrect data")
+		return
+	}
+
 	events, err := hdlr.service.GetEvents(periodParams)
 	if err != nil {
 		throwIntServerErr(wtr)
@@ -62,6 +80,16 @@ func (hdlr *handler) UpdateEvent(wtr http.ResponseWriter, req *http.Request) {
 
 	var eventData dto.Event
 	json.NewDecoder(req.Body).Decode(&eventData)
+
+	if eventData.Id == "" ||
+		!isDateValid(eventData.Date) ||
+		!isTimeValid(eventData.StartTime) ||
+		!isTimeValid(eventData.EndTime) ||
+		eventData.StartTime > eventData.EndTime ||
+		eventData.Content == "" {
+		throwBadReqErr(wtr, "Incorrect data")
+		return
+	}
 
 	err = hdlr.service.UpdateEvent(eventData)
 	if err != nil {
@@ -83,6 +111,11 @@ func (hdlr *handler) DeleteEvent(wtr http.ResponseWriter, req *http.Request) {
 
 	var eventData dto.Event
 	json.NewDecoder(req.Body).Decode(&eventData)
+
+	if eventData.Id == "" {
+		throwBadReqErr(wtr, "Incorrect data")
+		return
+	}
 
 	err = hdlr.service.DeleteEvent(eventData.Id)
 	if err != nil {
